# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IOMTea API",
    "version": "2.0.0",
    "description": "Home health IoT monitoring"
  },
...
  "paths": {
...
    "/pins": {
      "get": {
        "tags": [
          "Pins"
        ],
        "responses": {
          "200": {
            "description": "PIN list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "pin": {
                        "type": "string"
                      },
                      "userId": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "label": {
                        "type": "string",
                        "nullable": true
                      },
                      "isVirtual": {
                        "type": "boolean",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "nullable": true
                      }
                    },
                    "required": [
                      "pin",
                      "userId",
                      "type",
                      "label",
                      "isVirtual",
                      "createdAt"
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pins"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pin": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 6,
                    "example": "123456"
                  },
                  "userId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "device",
                      "virtual",
                      "user",
                      "simulator"
                    ],
                    "default": "device"
                  },
                  "label": {
                    "type": "string"
                  }
                },
                "required": [
                  "userId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pin": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string",
                      "nullable": true
                    },
                    "isVirtual": {
                      "type": "boolean",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "pin",
                    "userId",
                    "type",
                    "label",
                    "isVirtual",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/tags": {
      "get": {
        "tags": [
          "Tags"
        ],
        "responses": {
          "200": {
            "description": "Tag list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "color": {
                        "type": "string",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "color",
                      "createdAt"
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tags"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50,
                    "example": "高血压"
                  },
                  "color": {
                    "type": "string",
                    "example": "#ff6b6b"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "color": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "color",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/patients": {
      "get": {
        "tags": [
          "Patients"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Patient list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "gender": {
                        "type": "string",
                        "nullable": true
                      },
                      "birthDate": {
                        "type": "string",
                        "nullable": true
                      },
                      "heightCm": {
                        "type": "number",
                        "nullable": true
                      },
                      "weightKg": {
                        "type": "number",
                        "nullable": true
                      },
                      "bloodType": {
                        "type": "string",
                        "nullable": true
                      },
                      "phone": {
                        "type": "string",
                        "nullable": true
                      },
                      "address": {
                        "type": "string",
                        "nullable": true
                      },
                      "status": {
                        "type": "string"
                      },
                      "tags": {
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "gender",
                      "birthDate",
                      "heightCm",
                      "weightKg",
                      "bloodType",
                      "phone",
                      "address",
                      "status",
                      "createdAt"
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patients"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "张三"
                  },
                  "birthDate": {
                    "type": "string",
                    "example": "1960-01-01"
                  },
                  "gender": {
                    "type": "string",
                    "example": "male"
                  },
                  "heightCm": {
                    "type": "number"
                  },
                  "weightKg": {
                    "type": "number"
                  },
                  "bloodType": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    }
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "gender": {
                      "type": "string",
                      "nullable": true
                    },
                    "birthDate": {
                      "type": "string",
                      "nullable": true
                    },
                    "heightCm": {
                      "type": "number",
                      "nullable": true
                    },
                    "weightKg": {
                      "type": "number",
                      "nullable": true
                    },
                    "bloodType": {
                      "type": "string",
                      "nullable": true
                    },
                    "phone": {
                      "type": "string",
                      "nullable": true
                    },
                    "address": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string"
                    },
                    "tags": {
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "gender",
                    "birthDate",
                    "heightCm",
                    "weightKg",
                    "bloodType",
                    "phone",
                    "address",
                    "status",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "patientId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "severity",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Alert list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "patientId": {
                        "type": "string"
                      },
                      "kind": {
                        "type": "string"
                      },
                      "metric": {
                        "type": "string"
                      },
                      "value": {
                        "nullable": true
                      },
                      "unit": {
                        "type": "string",
                        "nullable": true
                      },
                      "severity": {
                        "type": "string",
                        "nullable": true
                      },
                      "status": {
                        "type": "string",
                        "nullable": true
                      },
                      "source": {
                        "type": "string",
                        "nullable": true
                      },
                      "recordedAt": {
                        "type": "string",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "nullable": true
                      },
                      "tags": {
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "patientId",
                      "kind",
                      "metric",
                      "unit",
                      "severity",
                      "status",
                      "source",
                      "recordedAt",
                      "createdAt"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}